+++ src/gwt/src/org/rstudio/studio/client/application/StudioClientApplicationConstants.java
@@ -202,4 +202,6 @@ public interface StudioClientApplicationConstants extends com.google.gwt.i18n.cl
     String workbenchLoginRequiredMessage();
     String serverLoginRequiredMessage();
     String loginButton();
+    String gpuSettingChangedCaption();
+    String gpuSettingChangedMessage();
 }